292 - Currency optimizations
