+++ thenewboston/currencies/filters/total_amount_minted.py
@@ -1,0 +1,11 @@
+import django_filters
+
+from ..models import Currency
+
+
+class TotalAmountMintedFilter(django_filters.FilterSet):
+    currency = django_filters.NumberFilter(field_name='id', required=True)
+
+    class Meta:
+        model = Currency
+        fields = ('currency',)
